refactor(common): rename SystemStatusBanner to convention
- Rename SystemStatusBanner → AstrosSystemStatusBanner (Astros prefix
  marks it as local code; matches the project component-naming rule).
- Move .vue + .stories.ts into common/systemStatusBanner/ per the
  per-component-folder convention used by AstrosLayout, AstrosSearchSelect,
  and the field components.
- Test renamed in __tests__/ to match the new component name; the
  __tests__/ subfolder layout matches the existing project test
  convention (stores/__tests__/, components/<domain>/__tests__/).
- Update AstrosLayout consumer (only consumer; not exported via the
  common/index.ts barrel) to the new import path and tag name.

All 3 file moves detected as 100% renames by git. 46/46 tests pass
including the 6 banner tests under the new file name.
