+++ astros_vue/src/components/common/layout/AstrosLayout.vue
@@ -2,7 +2,7 @@
 import router from '@/router';
 import apiService from '@/api/apiService';
 import { ref, watch } from 'vue';
-import SystemStatusBanner from '@/components/common/SystemStatusBanner.vue';
+import AstrosSystemStatusBanner from '@/components/common/systemStatusBanner/AstrosSystemStatusBanner.vue';
 
 const props = defineProps({
   isSidebarOpen: {
@@ -45,7 +45,7 @@ function logout() {
       aria-controls="sidebar-menu"
     />
     <div class="drawer-content">
-      <SystemStatusBanner />
+      <AstrosSystemStatusBanner />
       <div class="navbar bg-base-100 shadow-sm">
         <div class="flex-none pl-2 flex items-center">
           <label
